New service for handling setting changes
[MAILPOET-4061]
