+++ mailpoet/tests/integration/API/JSON/v1/SettingsTest.php
@@ -9,7 +9,6 @@ use MailPoet\API\JSON\Response as APIResponse;
 use MailPoet\API\JSON\v1\Settings;
 use MailPoet\Config\ServicesChecker;
 use MailPoet\Cron\Workers\InactiveSubscribers;
-use MailPoet\Cron\Workers\WooCommerceSync;
 use MailPoet\Entities\NewsletterEntity;
 use MailPoet\Entities\ScheduledTaskEntity;
 use MailPoet\Form\FormMessageController;
@@ -21,6 +20,7 @@ use MailPoet\Newsletter\Sending\ScheduledTasksRepository;
 use MailPoet\Segments\SegmentsRepository;
 use MailPoet\Services\AuthorizedEmailsController;
 use MailPoet\Services\Bridge;
+use MailPoet\Settings\SettingsChangeHandler;
 use MailPoet\Settings\SettingsController;
 use MailPoet\Settings\SettingsRepository;
 use MailPoet\Settings\TrackingConfig;
@@ -40,16 +40,12 @@ class SettingsTest extends \MailPoetTest {
   /** @var SettingsController */
   private $settings;
 
-  /** @var ScheduledTasksRepository */
-  private $tasksRepository;
-
   /* @var NewslettersRepository */
   private $newsletterRepository;
 
   public function _before() {
     parent::_before();
     ORM::raw_execute('TRUNCATE ' . ScheduledTask::$_table);
-    $this->tasksRepository = $this->diContainer->get(ScheduledTasksRepository::class);
     $this->newsletterRepository = $this->diContainer->get(NewslettersRepository::class);
     $this->settings = SettingsController::getInstance();
     $this->settings->set('some.setting.key', true);
@@ -66,6 +62,7 @@ class SettingsTest extends \MailPoetTest {
       $this->diContainer->get(FormMessageController::class),
       $this->make(ServicesChecker::class, ['isMailPoetAPIKeyPendingApproval' => false]),
       $this->diContainer->get(SegmentsRepository::class),
+      $this->diContainer->get(SettingsChangeHandler::class),
       $this->diContainer->get(SubscribersCountsController::class),
       $this->diContainer->get(TrackingConfig::class)
     );
@@ -108,6 +105,7 @@ class SettingsTest extends \MailPoetTest {
       $this->diContainer->get(FormMessageController::class),
       $this->make(ServicesChecker::class, ['isMailPoetAPIKeyPendingApproval' => false]),
       $this->diContainer->get(SegmentsRepository::class),
+      $this->diContainer->get(SettingsChangeHandler::class),
       $this->diContainer->get(SubscribersCountsController::class),
       $this->diContainer->get(TrackingConfig::class)
     );
@@ -141,6 +139,7 @@ class SettingsTest extends \MailPoetTest {
       $this->diContainer->get(FormMessageController::class),
       $this->make(ServicesChecker::class, ['isMailPoetAPIKeyPendingApproval' => false]),
       $this->diContainer->get(SegmentsRepository::class),
+      $this->diContainer->get(SettingsChangeHandler::class),
       $this->diContainer->get(SubscribersCountsController::class),
       $this->diContainer->get(TrackingConfig::class)
     );
@@ -169,6 +168,7 @@ class SettingsTest extends \MailPoetTest {
       $this->diContainer->get(FormMessageController::class),
       $this->make(ServicesChecker::class, ['isMailPoetAPIKeyPendingApproval' => false]),
       $this->diContainer->get(SegmentsRepository::class),
+      $this->diContainer->get(SettingsChangeHandler::class),
       $this->diContainer->get(SubscribersCountsController::class),
       $this->diContainer->get(TrackingConfig::class)
     );
@@ -199,6 +199,7 @@ class SettingsTest extends \MailPoetTest {
       $this->diContainer->get(FormMessageController::class),
       $this->make(ServicesChecker::class),
       $this->diContainer->get(SegmentsRepository::class),
+      $this->diContainer->get(SettingsChangeHandler::class),
       $this->diContainer->get(SubscribersCountsController::class),
       $this->diContainer->get(TrackingConfig::class)
     );
@@ -247,54 +248,6 @@ class SettingsTest extends \MailPoetTest {
     expect($this->settings->get('reply_to'))->isEmpty();
   }
 
-  public function testItReschedulesScheduledTaskForWoocommerceSync(): void {
-    $newTask = $this->createScheduledTask(WooCommerceSync::TASK_TYPE);
-    assert($newTask instanceof ScheduledTaskEntity);
-
-    $this->endpoint->onSubscribeOldWoocommerceCustomersChange();
-
-    $this->entityManager->clear();
-    $task = $this->getScheduledTaskByType(WooCommerceSync::TASK_TYPE);
-    assert($task instanceof ScheduledTaskEntity);
-    $scheduledAt = $task->getScheduledAt();
-    assert($scheduledAt instanceof \DateTime);
-    $expectedScheduledAt = Carbon::createFromTimestamp(WPFunctions::get()->currentTime('timestamp'));
-    $expectedScheduledAt->subMinute();
-    expect($scheduledAt)->equals($expectedScheduledAt);
-    expect($newTask->getId())->equals($task->getId());
-  }
-
-  public function testItCreatesScheduledTaskForWoocommerceSync(): void {
-    $task = $this->getScheduledTaskByType(WooCommerceSync::TASK_TYPE);
-    expect($task)->null();
-    $this->endpoint->onSubscribeOldWoocommerceCustomersChange();
-    $task = $this->getScheduledTaskByType(WooCommerceSync::TASK_TYPE);
-    expect($task)->isInstanceOf(ScheduledTaskEntity::class);
-  }
-
-  public function testItReschedulesScheduledTaskForInactiveSubscribers(): void {
-    $newTask = $this->createScheduledTask(InactiveSubscribers::TASK_TYPE);
-    assert($newTask instanceof ScheduledTaskEntity);
-    $this->endpoint->onInactiveSubscribersIntervalChange();
-
-    $task = $this->getScheduledTaskByType(InactiveSubscribers::TASK_TYPE);
-    assert($task instanceof ScheduledTaskEntity);
-    $scheduledAt = $task->getScheduledAt();
-    assert($scheduledAt instanceof \DateTime);
-    $expectedScheduledAt = Carbon::createFromTimestamp(WPFunctions::get()->currentTime('timestamp'));
-    $expectedScheduledAt->subMinute();
-    expect($scheduledAt)->equals($expectedScheduledAt);
-    expect($newTask->getId())->equals($task->getId());
-  }
-
-  public function testItCreatesScheduledTaskForInactiveSubscribers(): void {
-    $task = $this->getScheduledTaskByType(InactiveSubscribers::TASK_TYPE);
-    expect($task)->null();
-    $this->endpoint->onInactiveSubscribersIntervalChange();
-    $task = $this->getScheduledTaskByType(InactiveSubscribers::TASK_TYPE);
-    expect($task)->isInstanceOf(ScheduledTaskEntity::class);
-  }
-
   public function testItDeactivatesReEngagementEmailsIfTrackingDisabled(): void {
     $this->createNewsletter(NewsletterEntity::TYPE_RE_ENGAGEMENT, NewsletterEntity::STATUS_ACTIVE);
     $this->settings->set('tracking', ['level' => TrackingConfig::LEVEL_PARTIAL]);
@@ -319,22 +272,6 @@ class SettingsTest extends \MailPoetTest {
     expect($response->meta['showNotice'])->equals(false);
     $response = $this->endpoint->set(['tracking' => ['level' => TrackingConfig::LEVEL_BASIC]]);
     expect($response->meta['showNotice'])->equals(false);
-  }
-
-  private function createScheduledTask(string $type): ScheduledTaskEntity {
-    $task = new ScheduledTaskEntity();
-    $task->setType($type);
-    $task->setStatus(ScheduledTaskEntity::STATUS_SCHEDULED);
-    $this->tasksRepository->persist($task);
-    $this->tasksRepository->flush();
-    return $task;
-  }
-
-  private function getScheduledTaskByType(string $type): ?ScheduledTaskEntity {
-    return $this->tasksRepository->findOneBy([
-      'type' => $type,
-      'status' => ScheduledTaskEntity::STATUS_SCHEDULED,
-    ]);
   }
 
   private function createNewsletter(string $type, string $status = NewsletterEntity::STATUS_DRAFT, $parent = null): NewsletterEntity {
